# "Fast" timelord logging fix, "slow" timelord process termination fix, and configurable "plot quality lookup" warning for harvesters