@thesemaphoreslim 2024-08-25
Added a check on the streamreaders to ensure they are initialized.  Should help pass the CI checks.